Auto-detect existing LXC setup
The following is a work in progress and has only been manually tested.
Essentially the idea goes, that if you already have an existing setup
and just want juju to take over, with the previsor that you already
have the certs and config setup correctly, then this just works.

The down side to this setup is two fold:

 1. We touch a LOT of surface area to get the certificates to be
 detected at every point (bootstrap, destroy, kill)
 1. The credentials are not stored correctly, as we have to touch
 destroy and kill commands, which feels wrong.

This is a demo, I'm unsure if we should continue, but it's interesting
that it works nicely using the lxc config.

Note: there are no tests, as I would like to get feedback before I
spend any more time on this.