Start importing remote applications
The following aims to just start the importing of remote applications,
even though I'm unsure if any of it is actually used i.e. subnets.

It may turn out we don't need to import the remoteApplications directly,
but instead import consuming model with the following:

 - ExternalController
 - RelationNetwork
 - FirewallRule
 - RemoteEntity

And then import offering model with the following:

 - Offer
 - OfferConnection
 - OfferAccessPermissions
 - RemoteEntity
 - RelationNetwork
 - FirewallRule