Update use of charm Strategy for AddCharm*.
Incorporate charm Strategy changes in the application and charms
facades.  This allows for implementation of download functionality for
CharmHub charms.

Parts of this change include using local interfaces instead of the
charmrepo one, to have more general methods for repositories.  It is not
a perfect abstration of a repository as only CharmHub requires
FindDownloadURL.