Download charmhub fixes
Unfortunately the http client expected the output of a request to be of
a type and the refresh api expected a slice of types, then the parsing
of the error messages became impossible.

Additional investigation of the specs around why the refresh api
integration tests fail is also sad. Currently I've skipped them, but the
charmhub api seems to be validating everything, so we're a bit stuck
atm.