Ensure we handle missing metadata
Metadata with in a charm is a required yaml configuration. When it's
missing from a charm, we can no longer treat the charm as a valid charm.
This should have been caught during a validation process on the store
side, so we should never get here. Yet we do, so we should handle that
sceanrio.