Initial idea for refresher API
The idea is that we can break the refreshing of a charm (previously
called upgrade charm) into distinict functions. Those functions can act
in a similar vein to the deploy command.

By passing in a larger config to each refreshing task, it can pick what
it requires in a smaller unit config and work on just that. This should
give rise to better test integration.