Default architecture model config
Implementing this default-architecture as a separate work so that we can
better review it as a break off unit.

In order to correctly understand what to deploy using what architecture
we really need to ask the instance it's going to be placed on, "what
architecture do you support?". The problem is that Juju doesn't model it
like this, so we throw things at a wall and hope the stick. In the way
we're throwing a charm with an architecture at a wall and hope it
sticks. Introducing a model-config (model-defaults etc) allows the user
to say if you don't pass one during deploy time, at least fallback to
one they can controller.

Why not use the controller architecture?

It will probably be wrong and when it's wrong it'll be wrong in a big
way. Most of the time if you do have a heterogeneous setup your
controllers will likely be different than the homogeneous work units.