Pass objectstore to destroy
This is a bit of pain, as the entities act like an ORM around the
underlying state types. So to actually add the objectstore dependency
we need to push it in when we remove and destory those entites. This
was masked in the pass by just creating a blob store on the fly with
the state that was around. We can no longer do that as the state might
actually be an s3 or file based setup. These are managed as worker
resources so it's not like we can fallback to the singleton pattern.
Workers need to be able to restart and this is just us owning that
responsibility.