Objectstore with factory for multiple namespaces
The following refactors the object store to allow access to both
the controller and model namespaces in one request. This is required
because we need the for hosted models when doing agent tools. We could
just pass the object store getter into the context, but that weakens
the model. Instead we are going to be explicit about what you can
access from the facade context. This does reduce flexibility, but
that's on purpose.