S3Client PUT and DELETE methods
Adds PUT and DELETE methods to the s3caller client. The s3caller
client is a wrapper around the s3client to ensure it works with
the dependency engine.

The code adds simple PutObject and DeleteObject methods to the
s3client. There is a new credentials type that allows us to swap
the credentials based on what the caller wants. This is going to be
interesting to manage, as you'll likely want a s3client that has
different credentials.

For now I've baked in the annonymous credentials and we can iterate
on that.