Allow the creation of a credentials based client.
The credential based client will require the watching of the
controller config. The credential based client will be pre-seeded
with the current credentials, any changes will then force the
update. There might be some delay getting the newly created
credentials (writing and then flushing through all the watchers),
but a retry mechanism will be needed on the other end to request
the client again.

To think of it, it might be better to pass in a func that we only
allow you to access a client from with in. Thus allowing us
to recall the credentials without otherworks spinning up and down.