feat: static analysis assert checks
We keep getting bitten by running asserts in transactions. If an
assert is run it prevents a retry. This makes the tests flakey.

The solution is to create tooling to prevent this from happening.

I've fixed about 10 tests like this in the past couple of weeks, so
this tools does an exhaustive search. We can then add it to our
CI to prevent it from landing.