chore: fix apiserver object tests and remove AddLocalCharm integration test
This commit fixes 3 tests from the objects suite, caused by a change in
the minimum hash prefix length.

It also removes the TestAddLocalCharmSuccess test from the
clientMacaroonIntegrationSuite (which uses testing.Factory). The reason
is that since it's an integration test, in order for it to fully work we
would need an actual charm object store that created the inputs in the
corresponding tables, whilst today we're using a noopobjectstore. This
test is indeed important and will be reinstated later.