feat: adds logsink services
In order for the logsink worker to be used as a dependency to the
domain services, we have to employ the usual trick of having it's
own set of services that allows the worker to interact with the
database with minimal set of methods. This is the same pattern
we've used with the provider tracker and the objectstore.

The logsink now has less dependencies and we don't need to access the
model in the apiserver everytime an agent requests the logsink
for a model. This is cached at the logsink worker level. It inverts
the dependency...

If in the future we ever want to remove the naming scheme on the
log files, it's easy as it's all encapsulated.