feat: remote blob retriever
This gets a blob from the remote api caller mesh network. It will
retry if there is no connection, otherwise it will fail if it can't
find the blob. It will race to get the blob trying all the remotes
on the mesh until it can no longer locate one.

The logic is very hairy because we don't want to cancel a context
of a the winning task because it needs to read that information off
the wire then cancel it. We could read it to a buffer first, which
would simplify a lot of the logic, yet we'd be writing to two tmp
files just to make the download logic simpler, and probably not by
much.