feat: model info from controller state
We can't use the model state for accessing the model info, because
when we create a model, the model info hasn't been created before
we try and access the domain services. Luckily, all of the information
is available to us on the controller. There is a v_model view that
allows us access that information, but again it's guarded by an
activation check. Having an unactivated view allows us to get the
data in the correct fashion. We'll most likely not want to do this
very often, but there are some use cases for the infrastructure we're
trying to put up.