refactor: possible deadlock in sshserver
There is a deadlock in the sshserver when shutting down. If there have
been no accepts sent to the server, then we put up an ssh server and
it's unable to shut down. As fix is *only* for test code, we can
move it to test code only by creating a config option for it. This
means in the normal production code we don't have to know about a test
artifact.

This also cleans up a leaky abstraction.