feat: add service remote application offerer
This adds the service methods for adding a remote application offerer.
The application name, offer UUID and various parts are validated to
ensure correctness. More importantly, we create a synthetic charm that
will be used for relations. The charm will be bare bones and just have
the required elements filled in.

By adding an application and a charm into the database using the
existing infrastructure, is that we don't need to enforce uniqueness
of an application/charm using other means than what we've currently got
in place.