feat: implement GetOffersWithConnections in domain
Provides extra data, OfferConnections, that GetOffers does not provide.
There is no need for all data to be in one transaction. In fact it
cannot be as some is in a model db and some in the controller db.
Therefore add onto the GetOffers service method, rather than
complicating it further.