feat: remove juju-db
These are the remnants of the juju-db snap. This was ok as it was never
used and was essentially dead code, but the upgrade steps now we've
moved to 4.0.0 would try and install the juju-db for it. Except there
will never be a 4.0 of juju-db. It's gone.

There rest of the code mentions MongoInfo and we no longer require
that.