feat: rework model import to handle deletion
Updates model import to correctly handle deletion. We don't want to
delete model until instructed. So we need to rejig a lot of the
internals for importing a model.

This was never done previously because the state package was driving
this. Now that we don't have state we have to do a lot of this
ourselves.